requirements.txt
biopython==1.79
requests==2.26.0